ready for numpy 2.0

+++ dedupe/convenience.py
@@ -43,7 +43,7 @@ def randomPairs(n_records: int, sample_size: int) -> IndicesIterator:
     else:
         try:
             random_pairs = numpy.array(
-                random.sample(range(n), sample_size), dtype=numpy.uint
+                random.sample(range(n), sample_size), dtype=numpy.uint64
             )
         except OverflowError:
             return randomPairsWithReplacement(n_records, sample_size)
@@ -51,10 +51,10 @@ def randomPairs(n_records: int, sample_size: int) -> IndicesIterator:
     b: int = 1 - 2 * n_records
 
     i = (-b - 2 * numpy.sqrt(2 * (n - random_pairs) + 0.25)) // 2
-    i = i.astype(numpy.uint)
+    i = i.astype(numpy.int64)
 
     j = random_pairs + i * (b + i + 2) // 2 + 1
-    j = j.astype(numpy.uint)
+    j = j.astype(numpy.uint64)
 
     return zip(i, j)
